requirements.txt
beautifulsoup4==4.3.2
nltk==3.0.1
pymongo==2.7.2
pytz
flask_wtf
python-dateutil
wtforms
flask_pymongo
flask_mail
Flask-Assets
flask-debugtoolbar
